requirements.txt
# Flask and Web Framework Dependencies
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
itsdangerous==2.1.2
blinker==1.7.0

# Database
SQLAlchemy==2.0.23
PyMySQL==1.1.0
cryptography==41.0.8

# AI and Language Processing
langchain==0.1.0
langchain-google-genai==1.0.1
langchain-community==0.0.13
langchain-core==0.1.15
google-generativeai==0.3.2
google-ai-generativelanguage==0.4.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2
Pillow==10.1.0

# Authentication and Security
PyJWT==2.8.0
bcrypt==4.1.2
passlib==1.7.4

# HTTP and API
requests==2.31.0
httpx==0.25.2
urllib3==2.1.0
certifi==2023.11.17

# Environment and Configuration
python-dotenv==1.0.0
python-decouple==3.8

# Utilities
numpy==1.26.2
pandas==2.1.4
python-dateutil==2.8.2
pytz==2023.3

# Development and Testing
pytest==7.4.3
pytest-flask==1.3.0

# Production Server
gunicorn==21.2.0
gunicorn==21.2.0

# Logging and Monitoring
colorama==0.4.6